+++ CHANGELOG.md
@@ -15,12 +15,16 @@ Refactoring, CVC & NDEF support (WIP)
 
 * Minor optimisation: return card status on card select
 
+
 * Improve authentikey mgmt
     * In InitiateSecureChannel(), returns the coordx of authentikey to allow unambiguous recovery of authentiokey pubkey on client side
 
+
 * Use transient AES key for secure channel encryption (Transient object do not wear flash memory)
 
+
 * Use transient elliptic key for the ephemeral privkey used in secure channel
+
 
 * Add support for fixed CVC
   * When enabled, the unlock secret is fixed during the lifetime of the applet and set during applet installation using a value provided in install parameters. 
@@ -28,13 +32,18 @@ Refactoring, CVC & NDEF support (WIP)
   * The implementation of CVC is based on the same unlock secret mechanism and counter, the main difference is that CVC is fixed and provided externally, not generated randomly.
 add command to check unlock secret
 
+
 * Add command to check unlock secret:
   * INS 0x54
   * This function check a given unlock counter and unlock_code and check validity.
   * This is useful for a client application to confirm their ownership is valid: if the applet returns SW_INCORRECT_UNLOCK_CODE, this probably means the cached unlock_secret in the application is wrong.
 
 
+* Add APDU command to sign a transaction hash with the privkey for a given key slot
+  * INS: 0x5B
+  * This function is only available when slot status is 'unsealed'.
 
+  
 ## [0.1-0.2]
 
 * refactor card-setup: allows to read info when setup is not done (changes are not allowed)